src/example/lambda-py2-deps/requirements.txt
ruamel.yaml==0.15.40
